requirements.txt
streamlit
python-dotenv
requests
google-genai
